Fix failing case creating host only interface
